requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
datasets>=3.0,<4.0
sentence-transformers>=3.0,<6.0
faiss-cpu>=1.12,<2.0
numpy>=1.26,<3.0
